requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-WTF>=1.2.0
Flask-Login>=0.6.3
Flask-Migrate>=4.0.5
Werkzeug>=3.0.0
Pillow>=10.4.0
email-validator>=2.1.0
psycopg2-binary>=2.9.7
WTForms>=3.0.1
gunicorn>=21.2.0